+++ docs/news.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* a new method `open` which deprecates `initialize`
 
 ## [1.2.6] - 2021-03-17
 ### Added