+++ docs/news.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.3.0] - 2021-06-05
 ### Added
 * a new method `open` which deprecates `initialize`
 
@@ -83,7 +85,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0-alpha - 2003-12-10
 
-[Unreleased]: https://github.com/lualdap/lualdap/compare/v1.2.6...HEAD
+[Unreleased]: https://github.com/lualdap/lualdap/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/lualdap/lualdap/compare/v1.2.6...v1.3.0
 [1.2.6]: https://github.com/lualdap/lualdap/compare/v1.2.5...v1.2.6
 [1.2.5]: https://github.com/lualdap/lualdap/compare/v1.2.4...v1.2.5
 [1.2.4]: https://github.com/lualdap/lualdap/compare/v1.2.4-rc1...v1.2.4